+++ src/WebJobs.Script.WebHost/Management/InstanceManager.cs
@@ -38,7 +38,7 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Management
             IEnvironment environment, ILogger<InstanceManager> logger, IMetricsLogger metricsLogger, IMeshServiceClient meshServiceClient, IRunFromPackageHandler runFromPackageHandler,
             IPackageDownloadHandler packageDownloadHandler)
         {
-            _client = clientFactory.CreateClient() ?? throw new ArgumentNullException(nameof(clientFactory));
+            _client = clientFactory?.CreateClient() ?? throw new ArgumentNullException(nameof(clientFactory));
             _webHostEnvironment = webHostEnvironment ?? throw new ArgumentNullException(nameof(webHostEnvironment));
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
             _metricsLogger = metricsLogger;