+++ src/WebJobs.Script.WebHost/Management/LinuxSpecialization/ManagedIdentityTokenProvider.cs
@@ -16,7 +16,7 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Management.LinuxSpecialization
         private const string ApiVersion = "1.0";
 
         private readonly IEnvironment _environment;
-        private readonly HttpClient _httpClient;
+        private readonly IHttpClientFactory _httpClientFactory;
         private readonly IMetricsLogger _metricsLogger;
         private readonly ILogger<ManagedIdentityTokenProvider> _logger;
 
@@ -24,9 +24,9 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Management.LinuxSpecialization
             IMetricsLogger metricsLogger, ILogger<ManagedIdentityTokenProvider> logger)
         {
             _environment = environment;
-            _httpClient = httpClientFactory.CreateClient();
             _metricsLogger = metricsLogger;
             _logger = logger;
+            _httpClientFactory = httpClientFactory;
         }
 
         private string GetRunFromPackageIdentity()
@@ -102,9 +102,10 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.Management.LinuxSpecialization
                 {
                     using (var httpRequestMessage = new HttpRequestMessage(HttpMethod.Get, tokenEndpoint))
                     {
+                        var httpClient = _httpClientFactory.CreateClient();
                         var msiToken = _environment.GetEnvironmentVariable(EnvironmentSettingNames.MsiSecret);
                         httpRequestMessage.Headers.Add(ScriptConstants.XIdentityHeader, msiToken);
-                        using (var response = await _httpClient.SendAsync(httpRequestMessage))
+                        using (var response = await httpClient.SendAsync(httpRequestMessage))
                         {
                             response.EnsureSuccessStatusCode();
                             var readAsStringAsync = await response.Content.ReadAsStringAsync();