+++ release_notes.md
@@ -3,6 +3,7 @@
 - My change description (#PR)
 -->
 - Updated PowerShell Worker (PS7) to [3.0.1045](https://github.com/Azure/azure-functions-powershell-worker/releases/tag/v3.0.1045)
+- Breaking change - removed HttpClient as a registered service and replaced with IHttpClientFactory (#7692)
 
 **Release sprint:** Sprint 107
 [ [bugs](https://github.com/Azure/azure-functions-host/issues?q=is%3Aissue+milestone%3A%22Functions+Sprint+107%22+label%3Abug+is%3Aclosed) | [features](https://github.com/Azure/azure-functions-host/issues?q=is%3Aissue+milestone%3A%22Functions+Sprint+107%22+label%3Afeature+is%3Aclosed) ]